JarvisAlive/requirements.txt
# HeyJarvis Core Dependencies
# Python 3.11+ required

# Core framework
langgraph==0.2.16
langchain==0.2.16
langchain-anthropic==0.1.23
anthropic==0.56.0

# OpenAI for DALL-E logo generation
openai==1.54.3

# Data validation and models
pydantic==2.5.2
semantic-version==2.10.0

# Redis for session storage
redis==5.0.1

# Rich console interface
rich==13.7.0

# Docker for sandboxing
docker==7.0.0

# Environment configuration
python-dotenv==1.0.0

# HTTP clients
aiohttp==3.9.1
requests==2.31.0

# Async utilities
async-timeout==4.0.3

# Retry logic
tenacity==8.2.3

# Date/time handling
python-dateutil==2.8.2

# Token counting for LLM context management
tiktoken==0.5.2

# JSON handling
orjson==3.9.15

# Template engine
jinja2==3.1.4
beautifulsoup4==4.12.3

# Utilities
uuid==1.30
pathlib2==2.3.7

# Redis async support for message bus
redis[hiredis]==5.0.1

# Process monitoring for agent health
psutil==5.9.8

# Statistics for analytics
statistics

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1